refactoring backdrop.dart to smaller files
Closes https://github.com/fluttercommunity/backdrop/issues/25
